DIBUJOS

+++ source/Lase.hx
@@ -1,5 +1,6 @@
 package;
 
+import flash.desktop.Clipboard;
 import flixel.FlxSprite;
 import flixel.system.FlxAssets.FlxGraphicAsset;
 import flixel.FlxG;
@@ -11,19 +12,24 @@ import flixel.FlxG;
 class Lase extends FlxSprite
 {
 	private var _time:Float = 0;
+	private var _time2:Float = 0;
+	private var _intervalo:Float = 0;
 	private var _timeCamDir:Float;
 	private var _recta:Bool;
 	private var _positivo:Bool;
+	private var _pega:Bool = true;
 	
-	public function new(?X:Float=0, ?Y:Float=0, ?recta:Bool, ?positivo:Bool, ?velocidad:Float, ?timeCamDir:Float, ?SimpleGraphic:FlxGraphicAsset) 
+	public function new(?X:Float=0, ?Y:Float=0, ?recta:Bool, ?positivo:Bool, ?velocidad:Float, ?timeCamDir:Float, ?intervalo:Float, ?SimpleGraphic:FlxGraphicAsset) 
 	{
 		super(X, Y, SimpleGraphic);
 		_timeCamDir = timeCamDir;
 		_recta = recta;
+		_intervalo = intervalo;
 		_positivo = positivo;
 		if (_recta == true)
 		{
 			makeGraphic(800, 5);
+			loadGraphic("assets/images/Lacer/lacer.png");
 			if (_positivo == true)
 			{
 				velocity.y = velocidad;
@@ -37,6 +43,7 @@ class Lase extends FlxSprite
 		else
 		{
 			makeGraphic(5, 600);
+			loadGraphic("assets/images/Lacer/lacerL.png");
 			if (_positivo == true)
 			{
 				velocity.x = velocidad;
@@ -46,7 +53,7 @@ class Lase extends FlxSprite
 				velocity.x = -velocidad;
 			}
 		}
-		color = 0x00FFFF00;
+		
 		FlxG.state.add(this);
 	}
 	
@@ -54,6 +61,7 @@ class Lase extends FlxSprite
 	{
 		super.update(elapsed);
 		_time += elapsed;
+		_time2 += elapsed;
 		if (_time > _timeCamDir)
 		{
 			if (_recta == true)
@@ -83,7 +91,35 @@ class Lase extends FlxSprite
 				}
 			}
 		}
-		if (FlxG.overlap(Reg.player, this))
+		if (_time2 > _intervalo && _pega == true)
+		{
+			_intervalo = _intervalo * 2;
+			_pega = false;
+			if (_recta == true)
+			{
+				loadGraphic("assets/images/Lacer/lacer_2.png");
+			}
+			else
+			{
+				loadGraphic("assets/images/Lacer/lacerL_2.png");
+			}
+			
+		}
+		if (_time2 > _intervalo && _pega == false)
+		{
+			_intervalo = _intervalo / 2;	
+			_pega = true;
+			_time2 = 0;
+			if (_recta == true)
+			{
+				loadGraphic("assets/images/Lacer/lacer.png");
+			}
+			else
+			{
+				loadGraphic("assets/images/Lacer/lacerL.png");
+			}
+		}
+		if (FlxG.overlap(Reg.player, this) && _pega == true)
 		{
 			Reg.player.kill();
 		}
